+++ src/components/CreateAccount/InputPassword.jsx
@@ -1,10 +1,11 @@
 import React from 'react'
 import i18n from '../../i18n'
 import PropTypes from 'prop-types'
+import Checkbox from '../Widgets/Checkbox'
+import ValidatedField from '../Widgets/ValidatedField'
 import './InputPassword.scss'
 
 class InputPassword extends React.Component {
-  
   static displayName = 'InputPassword'
 
   static propTypes = {
@@ -14,7 +15,7 @@ class InputPassword extends React.Component {
     /** handles value changes */
     onChange: PropTypes.func.isRequired,
     /** value to initialize the input field */
-    value: PropTypes.string,
+    value: PropTypes.string.isRequired,
     /** true: input is visible as plaintext */
     show: PropTypes.bool
   }
@@ -22,7 +23,7 @@ class InputPassword extends React.Component {
   static defaultProps = {
     value: '',
     placeholder: '',
-    show: false,
+    show: false
   }
   constructor(props) {
     super(props)
@@ -32,31 +33,30 @@ class InputPassword extends React.Component {
   }
   render() {
     return (
-    <div className="input-password">
-      <input
-        autoFocus
-        type={this.state.showPassword ? 'text' : 'password'}
-        placeholder={this.props.placeholder}
-        className={this.props.className}
-        value={this.props.value}
-        onChange={this.props.onChange}
-      />
-      <div className="show-password-container">
-        <input
-          id="show-password-checkbox"
-          type="checkbox"
-          name="elements_input_bool"
-          className="show-password"
-          checked={this.state.showPassword}
-          onChange={() =>
-            this.setState({ showPassword: !this.state.showPassword })
-          }
+      <div className="input-password">
+        <ValidatedField
+          autoFocus
+          type={this.state.showPassword ? 'text' : 'password'}
+          placeholder={this.props.placeholder}
+          className={this.props.className}
+          value={this.props.value}
+          onChange={this.props.onChange}
         />
-        <label htmlFor="show-password-checkbox">
-          {i18n.t('mist.popupWindows.importAccount.buttons.showPassword')}
-        </label>
+        <div className="show-password-container">
+          <Checkbox
+            id="show-password-checkbox"
+            name="elements_input_bool"
+            className="show-password"
+            checked={this.state.showPassword}
+            onChange={() =>
+              this.setState({ showPassword: !this.state.showPassword })
+            }
+            labelText={i18n.t(
+              'mist.popupWindows.importAccount.buttons.showPassword'
+            )}
+          />
+        </div>
       </div>
-    </div>
     )
   }
 }