+++ CHANGELOG.md
@@ -2,6 +2,106 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="1.1.6"></a>
+## [1.1.6](https://github.com/PhilippLgh/ethereum-react-components/compare/v1.1.5...v1.1.6) (2018-12-03)
+
+
+### Bug Fixes
+
+* some sanity checks, title tooltip, added passable classes et al ([#28](https://github.com/PhilippLgh/ethereum-react-components/issues/28)) ([61372ba](https://github.com/PhilippLgh/ethereum-react-components/commit/61372ba))
+
+
+
+<a name="1.1.5"></a>
+## [1.1.5](https://github.com/PhilippLgh/ethereum-react-components/compare/v1.1.4...v1.1.5) (2018-12-03)
+
+
+### Bug Fixes
+
+* sanitycheck for missing props and adding custom classes ([#29](https://github.com/PhilippLgh/ethereum-react-components/issues/29)) ([fe2afd9](https://github.com/PhilippLgh/ethereum-react-components/commit/fe2afd9))
+
+
+
+<a name="1.1.4"></a>
+## [1.1.4](https://github.com/PhilippLgh/ethereum-react-components/compare/v1.1.3...v1.1.4) (2018-11-30)
+
+
+### Bug Fixes
+
+* augments FormSubmitTx and Button ([#26](https://github.com/PhilippLgh/ethereum-react-components/issues/26)) ([cbfed7d](https://github.com/PhilippLgh/ethereum-react-components/commit/cbfed7d))
+
+
+
+<a name="1.1.3"></a>
+## [1.1.3](https://github.com/PhilippLgh/ethereum-react-components/compare/v1.1.2...v1.1.3) (2018-11-30)
+
+
+### Bug Fixes
+
+* updates EthAddress with wallet feedback ([#25](https://github.com/PhilippLgh/ethereum-react-components/issues/25)) ([bc6940e](https://github.com/PhilippLgh/ethereum-react-components/commit/bc6940e))
+
+
+
+<a name="1.1.2"></a>
+## [1.1.2](https://github.com/PhilippLgh/ethereum-react-components/compare/v1.1.1...v1.1.2) (2018-11-30)
+
+
+### Bug Fixes
+
+* introduces Button styling with variations ([#24](https://github.com/PhilippLgh/ethereum-react-components/issues/24)) ([1d2a03a](https://github.com/PhilippLgh/ethereum-react-components/commit/1d2a03a))
+
+
+
+<a name="1.1.1"></a>
+## [1.1.1](https://github.com/PhilippLgh/ethereum-react-components/compare/v1.1.0...v1.1.1) (2018-11-28)
+
+
+### Bug Fixes
+
+* brings Identicon to parity with Mist ([#21](https://github.com/PhilippLgh/ethereum-react-components/issues/21)) ([b3ade0e](https://github.com/PhilippLgh/ethereum-react-components/commit/b3ade0e))
+
+
+
+<a name="1.1.0"></a>
+# [1.1.0](https://github.com/PhilippLgh/ethereum-react-components/compare/v1.0.8...v1.1.0) (2018-11-27)
+
+
+### Features
+
+* Introduces ValidatedField within InputPassword ([#20](https://github.com/PhilippLgh/ethereum-react-components/issues/20)) ([87c17fc](https://github.com/PhilippLgh/ethereum-react-components/commit/87c17fc))
+
+
+
+<a name="1.0.8"></a>
+## [1.0.8](https://github.com/PhilippLgh/ethereum-react-components/compare/v1.0.7...v1.0.8) (2018-11-26)
+
+
+
+<a name="1.0.7"></a>
+## [1.0.7](https://github.com/PhilippLgh/ethereum-react-components/compare/v1.0.6...v1.0.7) (2018-11-26)
+
+
+
+<a name="1.0.6"></a>
+## [1.0.6](https://github.com/PhilippLgh/ethereum-react-components/compare/v1.0.5...v1.0.6) (2018-11-26)
+
+
+
+<a name="1.0.5"></a>
+## [1.0.5](https://github.com/PhilippLgh/ethereum-react-components/compare/v1.0.4...v1.0.5) (2018-11-23)
+
+
+
+<a name="1.0.4"></a>
+## [1.0.4](https://github.com/PhilippLgh/ethereum-react-components/compare/v1.0.3...v1.0.4) (2018-11-22)
+
+
+
+<a name="1.0.3"></a>
+## [1.0.3](https://github.com/PhilippLgh/ethereum-react-components/compare/v1.0.2...v1.0.3) (2018-11-21)
+
+
+
 <a name="1.0.2"></a>
 ## [1.0.2](https://github.com/PhilippLgh/ethereum-react-components/compare/v1.0.1...v1.0.2) (2018-11-21)